Moved leaderboard tests into separate file, deleted big tests.py file
